requirements.txt
alembic==1.16.1
APScheduler==3.10.4
beautifulsoup4==4.12.3
httpx==0.28.1
lxml==5.3.0
psycopg[binary]==3.2.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
SQLAlchemy==2.0.41
pytest==8.3.5
tzdata==2025.2
